rename Reveal to Inquire
