+++ package.json
@@ -302,13 +302,13 @@
         "category": "Zowe Explorer for IBM CICS"
       },
       {
-        "command": "cics-extension-for-zowe.revealTransaction",
-        "title": "Reveal Transaction",
+        "command": "cics-extension-for-zowe.inquireTransaction",
+        "title": "Inquire Transaction",
         "category": "Zowe Explorer for IBM CICS"
       },
       {
-        "command": "cics-extension-for-zowe.revealProgram",
-        "title": "Reveal Program",
+        "command": "cics-extension-for-zowe.inquireProgram",
+        "title": "Inquire Program",
         "category": "Zowe Explorer for IBM CICS"
       },
       {
@@ -548,13 +548,13 @@
         },
         {
           "when": "view == cics-view && viewItem =~ /^cicstask.*/",
-          "command": "cics-extension-for-zowe.revealTransaction",
-          "group": "001_zowecics_reveal@0"
+          "command": "cics-extension-for-zowe.inquireTransaction",
+          "group": "001_zowecics_inquire@0"
         },
         {
           "when": "view == cics-view && viewItem =~ /^cicstransaction.*/",
-          "command": "cics-extension-for-zowe.revealProgram",
-          "group": "001_zowecics_reveal@0"
+          "command": "cics-extension-for-zowe.inquireProgram",
+          "group": "001_zowecics_inquire@0"
         },
         {
           "when": "view == cics-view && viewItem =~ /^cicstask.*/",
