+++ src/commands/inquireTransaction.ts
@@ -9,30 +9,30 @@
 *
 */
 
-import { commands, TreeView, window } from "vscode";
-import { CICSCombinedTransactionsTree } from "../trees/CICSCombinedTransactionTree";
+import { commands, TreeItemCollapsibleState, TreeView, window } from "vscode";
+import { CICSCombinedTaskTree } from "../trees/CICSCombinedTaskTree";
 import { CICSPlexTree } from "../trees/CICSPlexTree";
-import { CICSProgramTree } from "../trees/CICSProgramTree";
 import { CICSRegionsContainer } from "../trees/CICSRegionsContainer";
 import { CICSRegionTree } from "../trees/CICSRegionTree";
+import { CICSTransactionTree } from "../trees/CICSTransactionTree";
 import { CICSTree } from "../trees/CICSTree";
-import { CICSTransactionTreeItem } from "../trees/treeItems/CICSTransactionTreeItem";
+import { CICSTaskTreeItem } from "../trees/treeItems/CICSTaskTreeItem";
 import { findSelectedNodes } from "../utils/commandUtils";
 
 /**
- * Reveal the associated transaction tree item from a task tree item
+ * Inquire the associated transaction tree item from a task tree item
  */
-export function getRevealProgramCommand(tree: CICSTree,treeview: TreeView<any>) {
+export function getInquireTransactionCommand(tree: CICSTree,treeview: TreeView<any>) {
     return commands.registerCommand(
-      "cics-extension-for-zowe.revealProgram",
+      "cics-extension-for-zowe.inquireTransaction",
       async (node) => {
-        const allSelectedNodes = findSelectedNodes(treeview, CICSTransactionTreeItem, node) as CICSTransactionTreeItem[];
+        const allSelectedNodes = findSelectedNodes(treeview, CICSTaskTreeItem, node) as CICSTaskTreeItem[];
         if (!allSelectedNodes || !allSelectedNodes.length) {
-          window.showErrorMessage("No CICS Transaction selected");
+          window.showErrorMessage("No CICS Task selected");
           return;
         }
         let resourceFolders;
-        if (allSelectedNodes[0].getParent() instanceof CICSCombinedTransactionsTree) {
+        if (allSelectedNodes[0].getParent() instanceof CICSCombinedTaskTree) {
           const cicsPlex: CICSPlexTree = allSelectedNodes[0].getParent().getParent();
           const regionsContainer = cicsPlex.getChildren().filter(child => child instanceof CICSRegionsContainer)[0];
           //@ts-ignore
@@ -42,22 +42,22 @@ export function getRevealProgramCommand(tree: CICSTree,treeview: TreeView<any>) 
           resourceFolders = allSelectedNodes[0].parentRegion.getChildren()!;
         }
         let tranids = [];
-        for (const localTransactionTreeItem of allSelectedNodes) {
-            const transaction = localTransactionTreeItem.transaction;
-            tranids.push(transaction["program"]);
+        for (const localTaskTreeItem of allSelectedNodes) {
+            const task = localTaskTreeItem.task;
+            tranids.push(task["tranid"]);
         }
         // Comma separated filter
         const pattern = tranids.join(", ");
-        const programTree = resourceFolders.filter(child => child instanceof CICSProgramTree)[0] as CICSProgramTree;
-        programTree.setFilter(pattern);
-        await programTree.loadContents();
+        const localTransactionTree = resourceFolders.filter(child => child instanceof CICSTransactionTree)[0] as CICSTransactionTree;
+        localTransactionTree.setFilter(pattern);
+        await localTransactionTree.loadContents();
         tree._onDidChangeTreeData.fire(undefined);
 
-        if (allSelectedNodes[0].getParent() instanceof CICSCombinedTransactionsTree) {
-          let nodeToExpand: any = programTree;
+        if (allSelectedNodes[0].getParent() instanceof CICSCombinedTaskTree) {
+          let nodeToExpand: any = localTransactionTree;
           
           // TODO: Ideal situation would be to do await treeview.reveal(nodeToExpand), however this is resulting in an error,
-          // so reveal/highlight parent node instead
+          // so inquire/highlight parent node instead
           await treeview.reveal(nodeToExpand.getParent());
           tree._onDidChangeTreeData.fire(undefined);
